requirements.txt
json
requests
beautifulsoup4
psycopg2-binary
python-dotenv
postgresql-client